Create a mixin for list-accumulator
